+++ packages/std/__tests__/types/isJSON.spec.ts
@@ -10,20 +10,21 @@ describe('Types > isJSON', () => {
     expect(isJSON([])).toBeFalsy();
     expect(isJSON('')).toBeFalsy();
     expect(isJSON('undefined')).toBeFalsy();
-    expect(isJSON('null')).toBeFalsy();
-    expect(isJSON('true')).toBeFalsy();
-    expect(isJSON('false')).toBeFalsy();
-    expect(isJSON('1')).toBeFalsy();
-    expect(isJSON('12345')).toBeFalsy();
     expect(isJSON('abc')).toBeFalsy();
     expect(isJSON('{a}')).toBeFalsy();
     expect(isJSON('{a:}')).toBeFalsy();
     expect(isJSON('{a:1}')).toBeFalsy();
     expect(isJSON('{"a": undefined}')).toBeFalsy();
-    expect(isJSON('[]')).toBeFalsy();
   });
 
   test('valid cases', () => {
+    expect(isJSON('null')).toBeTruthy();
+    expect(isJSON('true')).toBeTruthy();
+    expect(isJSON('false')).toBeTruthy();
+    expect(isJSON('1')).toBeTruthy();
+    expect(isJSON('12345')).toBeTruthy();
+    expect(isJSON('[]')).toBeTruthy();
+    expect(isJSON('"hello"')).toBeTruthy();
     expect(isJSON('{}')).toBeTruthy();
     expect(isJSON('{"a":1}')).toBeTruthy();
     expect(